mostly working general pane
